Code Rush Based cleanup - Move files and types - Ignored CRR0052 - string compairs - multiple validation combinations
