+++ Autabee.Automation.Utility.IEC61131TypeConversionTests/TestObjects/Address.cs
@@ -1,0 +1,8 @@
+﻿namespace Autabee.Utility.IEC61131TypeConversionTests.TestObjects
+{
+    class Address
+    {
+        public string Postcode { get; set; }
+        public GpsCoordinates GPS { get; set; }
+    }
+}
