+++ Autabee.Automation.Utility.IEC61131TypeConversionTests/TestObjects/GpsCoordinates.cs
@@ -1,0 +1,8 @@
+﻿namespace Autabee.Utility.IEC61131TypeConversionTests.TestObjects
+{
+    class GpsCoordinates
+    {
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
+    }
+}
