+++ Autabee.Automation.Utility.IEC61131TypeConversionTests/TestObjects/Person.cs
@@ -1,0 +1,9 @@
+﻿namespace Autabee.Utility.IEC61131TypeConversionTests.TestObjects
+{
+    class Person
+    {
+        public string Name { get; set; }
+        public int PasswordHash { get; set; }
+        public Address Address { get; set; }
+    }
+}
