+++ Autabee.Utility/ValidationResult.cs
@@ -47,23 +47,4 @@ namespace Autabee.Utility
 
         public bool Success { get; private set; }
     }
-
-    public class ValidationResult<T> : ValidationResult
-    {
-        public ValidationResult(bool success = true, T Object = default, string unSuccesfullText = "", params object[] formatObjects) : base(success, unSuccesfullText, formatObjects)
-        {
-            this.Object = Object;
-        }
-
-        public ValidationResult(bool success = true, T Object = default, List<(string, object[])> failInfo = default) : base(success, failInfo)
-        {
-            this.Object = Object;
-        }
-
-        public ValidationResult(ValidationResult result) : base(result.Success, result.FailInfo)
-        {
-            this.Object = default;
-        }
-        public T Object { get; }
-    }
 }
