+++ Open.PackML/Interfaces/IPmlEventStore`1.cs
