+++ Open.PackMLTests/TestObjects/EventHanderEnum1.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Open.PackMLTests.TestObjects
+{
+    public enum EventHanderEnum1
+    {
+        id1,
+        id2,
+        id3
+    }
+}
