+++ Open.PackMLTests/TestObjects/EventHanderEnum2.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Open.PackMLTests.TestObjects
+{
+    public enum EventHanderEnum2
+    {
+        id1,
+        id2,
+        id3
+    }
+}
